Move `rapidjson` dump functions out of public API
