+++ src/json/gauge_json_writer.hpp
@@ -15,8 +15,8 @@
 * License along with this library.
 */
 
-#ifndef HANDYSTATS_TIMER_JSON_WRITER_HPP_
-#define HANDYSTATS_TIMER_JSON_WRITER_HPP_
+#ifndef HANDYSTATS_GAUGE_JSON_WRITER_HPP_
+#define HANDYSTATS_GAUGE_JSON_WRITER_HPP_
 
 #include <string>
 
@@ -24,13 +24,14 @@
 #include <rapidjson/stringbuffer.h>
 #include <rapidjson/prettywriter.h>
 
-#include <handystats/json/statistics_json_writer.hpp>
-#include <handystats/metrics/timer.hpp>
+#include <handystats/metrics/gauge.hpp>
+
+#include "json/statistics_json_writer.hpp"
 
 namespace handystats { namespace json {
 
 template<typename Allocator>
-inline void write_to_json_value(const metrics::timer* const obj, rapidjson::Value* json_value, Allocator& allocator) {
+inline void write_to_json_value(const metrics::gauge* const obj, rapidjson::Value* json_value, Allocator& allocator) {
 	if (!obj) {
 		json_value = new rapidjson::Value();
 		return;
@@ -43,13 +44,13 @@ inline void write_to_json_value(const metrics::timer* const obj, rapidjson::Valu
 		json_value->SetObject();
 	}
 
-	json_value->AddMember("type", "timer", allocator);
+	json_value->AddMember("type", "gauge", allocator);
 
 	write_to_json_value(&obj->values(), json_value, allocator);
 }
 
 template<typename StringBuffer, typename Allocator>
-inline void write_to_json_buffer(const metrics::timer* const obj, StringBuffer* buffer, Allocator& allocator) {
+inline void write_to_json_buffer(const metrics::gauge* const obj, StringBuffer* buffer, Allocator& allocator) {
 	rapidjson::Value json_value;
 	write_to_json_value(obj, &json_value, allocator);
 
@@ -62,7 +63,7 @@ inline void write_to_json_buffer(const metrics::timer* const obj, StringBuffer* 
 }
 
 template<typename Allocator>
-inline std::string write_to_json_string(const metrics::timer* const obj, Allocator&& allocator = Allocator()) {
+inline std::string write_to_json_string(const metrics::gauge* const obj, Allocator&& allocator = Allocator()) {
 	rapidjson::GenericStringBuffer<rapidjson::UTF8<>, Allocator> buffer(&allocator);
 	write_to_json_buffer(obj, &buffer, allocator);
 
@@ -71,4 +72,4 @@ inline std::string write_to_json_string(const metrics::timer* const obj, Allocat
 
 }} // namespace handystats::json
 
-#endif // HANDYSTATS_TIMER_JSON_WRITER_HPP_
+#endif // HANDYSTATS_GAUGE_JSON_WRITER_HPP_
