+++ src/json/statistics_json_writer.hpp
@@ -25,8 +25,9 @@
 #include <rapidjson/stringbuffer.h>
 #include <rapidjson/prettywriter.h>
 
-#include <handystats/json/timestamp.hpp>
 #include <handystats/statistics.hpp>
+
+#include "json/timestamp.hpp"
 
 namespace handystats { namespace json {
