+++ tests/counter.gtest.cpp
@@ -22,7 +22,7 @@
 #include <handystats/chrono.hpp>
 #include <handystats/metrics/counter.hpp>
 
-#include <handystats/json/counter_json_writer.hpp>
+#include "json/counter_json_writer.hpp"
 
 using namespace handystats::metrics;
